Adequando as responsabilidades de cada camada da aplicação
