+++ server/src/controller/userController.js
@@ -1,39 +1,17 @@
-const mongoose = require("mongoose");
-const UserSchema = require("../db/schema/userSchema");
-const validator = require("../utils/validators/userValidator");
+const userModel = require("../db/model/userModel");
 
-const UserModel = mongoose.model("user", UserSchema);
-
-const user = {
-	async listUsers(_id) {
-		if (_id) {
-			return UserModel.find({ _id: _id });
-		}
-		return UserModel.find({});
+const controller = {
+	async getUsers(req, res, _next) {
+		res.send(await userModel.listUsers(req.query.userId));
 	},
-	async insertUser(userArray) {
-		if (!validator.hasInvalidField(userArray)) {
-			try {
-				return UserModel.insertMany(userArray);
-			} catch (error) {
-				return {
-					statusCode: 500,
-					error: "internal error",
-				};
-			}
-		}
+	async newUser(req, res, next) {
+		res.send(await userModel.insertUser(req.body));
 	},
-	async deleteUser(cpf) {
-		try {
-			return UserModel.deleteOne({ cpf: cpf });
-		} catch (error) {
-			return {
-				statusCode: 500,
-				error: "internal error",
-			};
-		}
+	async updateUser(req, res, next) {
+		res.send(await userModel.updateUser(req.body));
 	},
-	UserModel,
+	async deleteUser(req, res, next) {
+		res.send(await userModel.deleteUser(req.query.userId));
+	},
 };
-
-module.exports = user;
+module.exports = controller;
