+++ server/src/db/model/userModel.js
@@ -1,0 +1,22 @@
+const mongoose = require("mongoose");
+const UserSchema = require("../schema/userSchema");
+
+const UserModel = mongoose.model("user", UserSchema);
+
+const user = {
+	async listUsers(_id) {
+		if (_id) {
+			return UserModel.find({ _id: _id });
+		}
+		return UserModel.find({});
+	},
+	async insertUser(userArray) {
+		return UserModel.insertMany(userArray);
+	},
+	async deleteUser(cpf) {
+		return UserModel.deleteOne({ cpf: cpf });
+	},
+	UserModel,
+};
+
+module.exports = user;
